Fix handling of UID=1000 in Dockerfile
We can't create a new user and assign it UID=1000, since that user already
exists: it's the `node` user created by our base package. Thus, we can just
reuse the user account, and if needed, set the UID to whatever is the UID of
the calling user, which simplifies our Dockerfile quite a bit.