Remove early prototype Go server using Soy
We won't be using Soy server-side rendering, because it doesn't seem to be
compatible with several open-source web frameworks we have tried, especially if
we plan to use TypeScript, which requires a build process, but JS/TS frameworks
don't understand Soy templates, typically only JSX/TSX.

We may revisit this in the future if/when Soy is well-supported by various web
frameworks; the only one we found (Metal.js) has been deprecated and is no
longer developed or supported, and the creators have moved on to using React.

For now, we'll use Preact with JSX, and statically render that to HTML/JS/CSS,
and then use a Go server to serve those shell pages, which means we'll be using
primarily client-side rendering, which may be slower, but will be much simpler
to implement, and will be using the supported development path today.