Add a proxy to mimic prod deployment
This is a simple Go server which sends requests to backends based on URL
patterns. Essentially, this is a much simpler version of Envoy, implementing a
tiny fraction of its capabilities with some standard libraries provided in Go.

This is enough to route incoming requests either to our web UI server (Node) or
the API server (another Go binary), and would mimic the actual production
deployment if this were done in App Engine or Cloud Run, behind a reverse
proxy which would direct requests similarly based on URL patterns.