Enable the web UI to call list and create APIs
On load, the UI now queries for all existing notes (the API server currently
returns a hard-coded list of several notes).

On save, the UI will send a POST request to create a new note, and display the
response from the server (the API server currently echoes back what it
received, along with a newly-created UID for the note).