Remove duplicate output for redirected URLs
When we are redirecting URLs, we are printing the URLs out twice:

* once for the incoming URL request
* once more for redirecting the URL, without any additional info

Replace this with a single line per handled request, and also output the method
with the URL to easily differentiate GET vs. POST requests, for example.

Also add the method to the error messages for unhandled URLs for clarity.